requirements.txt
Flask==1.1.2
pandas==1.0.5
psycopg2-binary==2.8.5
plotly==4.14.1
requests